backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
googletrans==4.0.0-rc1
sqlalchemy
